Update PageUserGlobal
* Change InstantMessage -> PageUser
* Fixed nested If statement
* Added >6 detection before larger IF statement
